Refactor of the UI.
