+++ FrontEnd/web/index.html
@@ -16,7 +16,7 @@
 
                 <label for="inputProjectName" class="sr-only">Project name</label>
                 <input type="text" id="inputProjectName" class="form-control" placeholder="Project name" required autofocus>
-                <br/>
+                <br />
 
                 <label for="inputSvnAdress" class="sr-only">SVN Andress (svn://host/project/trunk)</label>
                 <input type="url" id="inputSvnAdress" class="form-control" placeholder="SVN Andress" required autofocus>
@@ -24,10 +24,10 @@
 
                 <label for="inputUsersFile" class="sr-only">Users' file</label>
                 <input type="text" id="inputUsersFile" class="form-control" placeholder="Users' file" required>
-                <br/>
+                <br />
 
-                <button class="btn btn-lg btn-primary btn-block" type="submit">Run</button>
-                <br/>
+                <button id="btnSubmit" class="btn btn-lg btn-primary btn-block" type="submit">Run</button>
+                <br />
 
                 <textarea id="logtextarea" cols="40" rows="5"></textarea>
             </form>
@@ -35,24 +35,41 @@
 
     </div> <!-- /container -->
 
-<script src="assets/js/jquery-1.12.0.min.js" type="text/javascript"></script>
-<script type="text/javascript">
-    function addlog(message) {
-        var txt = $("#logtextarea");
-        txt.val(txt.val() + "\n" + message);
-    }
+    <script src="assets/js/jquery-1.12.0.min.js" type="text/javascript"></script>
+    <script type="text/javascript">
+        function addlog(message) {
+            var txt = $("#logtextarea");
+            txt.val(txt.val() + "\n" + message);
+        }
 
-    $(function() {
-        $("#formMain").submit(function (event) {
-            var projectName = $("#inputProjectName").val();
-            var svnAdress = $("#inputSvnAdress").val();
-            var usersFile = $("#inputUsersFile").val();
+        function success() {
+            var txt = $("#logtextarea");
+            txt.val(txt.val() + "\nSuccess!");
 
-            bound.execute(svnAdress, usersFile, projectName);
+            $("#btnSubmit").removeAttr("disabled");
+        }
 
-            event.preventDefault();
+        function error(message) {
+            var txt = $("#logtextarea");
+            txt.val(txt.val() + "\nError: " + message);
+
+            $("#btnSubmit").removeAttr("disabled");
+        }
+
+        $(function () {
+            $("#formMain").submit(function (event) {
+                var svnAdress = $("#inputSvnAdress").val();
+                var usersFile = $("#inputUsersFile").val();
+                var projectName = $("#inputProjectName").val();
+                
+                bound.execute(svnAdress, usersFile, projectName);
+
+                $("#btnSubmit").attr("disabled", true);
+                event.preventDefault();
+            });
+
+            $("#inputSvnAdress").focus();
         });
-    });
-</script>
+    </script>
 </body>
 </html>
